change structure files
