+++ srcs/requirements/Nginx/conf/nginxTranscend.conf
@@ -38,6 +38,16 @@ http {
             proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
             proxy_set_header X-Forwarded-Proto https;
             proxy_set_header Upgrade $http_upgrade;
+        }
+
+        location /ws/
+        {
+            proxy_pass http://daphne:8001;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto https;
+            proxy_set_header Upgrade $http_upgrade;
             proxy_set_header Connection "Upgrade";
         }
